# docs/api.openapi.yaml
openapi: 3.1.0
info:
  title: AI Image Studio API
  version: 0.1.0
  description: Flutter 客户端使用的 AI 图片服务接口。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      security: []
      summary: 注册并创建会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '409':
          $ref: '#/components/responses/Error'
  /auth/login:
    post:
      security: []
      summary: 登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Error'
  /auth/refresh:
    post:
      security: []
      summary: 轮换 Refresh Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
                device_name: {type: string}
      responses:
        '200':
          description: 新令牌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        '401':
          $ref: '#/components/responses/Error'
  /auth/logout:
    post:
      security: []
      summary: 撤销 Refresh Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: {type: string}
      responses:
        '204': {description: 已退出}
  /me:
    get:
      summary: 当前用户
      responses:
        '200':
          description: 用户资料和额度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Error'
  /styles:
    get:
      security: []
      summary: 可用风格预设
      responses:
        '200':
          description: 风格列表
          content:
            application/json:
              schema:
                type: object
                required: [styles]
                properties:
                  styles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Style'
  /assets:
    post:
      summary: 上传参考图片
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: 已保存资产
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '413':
          $ref: '#/components/responses/Error'
        '415':
          $ref: '#/components/responses/Error'
  /assets/{assetId}:
    parameters:
      - $ref: '#/components/parameters/AssetId'
    get:
      summary: 获取资产和临时访问 URL
      responses:
        '200':
          description: 图片资产
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      summary: 删除资产
      responses:
        '204': {description: 已删除}
        '404':
          $ref: '#/components/responses/Error'
  /image-tasks:
    post:
      summary: 创建图片生成任务
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageTask'
      responses:
        '202':
          description: 任务已入队
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageTask'
        '402':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
    get:
      summary: 作品历史
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: cursor
          in: query
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 50, default: 20}
      responses:
        '200':
          description: 游标分页任务
          content:
            application/json:
              schema:
                type: object
                required: [tasks, next_cursor]
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageTask'
                  next_cursor: {type: string}
  /image-tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      summary: 查询任务
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageTask'
        '404':
          $ref: '#/components/responses/Error'
  /image-tasks/{taskId}/cancel:
    post:
      summary: 取消尚未开始的任务并退款
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204': {description: 已取消并退款}
        '409':
          $ref: '#/components/responses/Error'
  /image-tasks/{taskId}/retry:
    post:
      summary: 重试失败或已取消的任务
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: 新任务已入队
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageTask'
        '409':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AssetId:
      name: assetId
      in: path
      required: true
      schema: {type: string}
    TaskId:
      name: taskId
      in: path
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, maxLength: 255}
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, nickname]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 8, maxLength: 72}
        nickname: {type: string, maxLength: 80}
        device_name: {type: string}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string}
        device_name: {type: string}
    User:
      type: object
      required: [id, email, nickname, credits_balance]
      properties:
        id: {type: string}
        email: {type: string, format: email}
        nickname: {type: string}
        avatar_asset_id: {type: string}
        credits_balance: {type: integer, format: int64}
    Tokens:
      type: object
      required: [access_token, refresh_token, token_type, access_expires_at, refresh_expires_at]
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        token_type: {type: string, const: Bearer}
        access_expires_at: {type: string, format: date-time}
        refresh_expires_at: {type: string, format: date-time}
    AuthSession:
      type: object
      required: [user, tokens]
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/Tokens'
    Style:
      type: object
      required: [id, slug, name, description]
      properties:
        id: {type: string}
        slug: {type: string}
        name: {type: string}
        description: {type: string}
    Asset:
      type: object
      required: [id, kind, url, mime_type, width, height, byte_size, ai_generated, created_at]
      properties:
        id: {type: string}
        kind: {type: string}
        url: {type: string, format: uri}
        thumbnail_url: {type: string, format: uri}
        mime_type: {type: string}
        width: {type: integer}
        height: {type: integer}
        byte_size: {type: integer, format: int64}
        ai_generated: {type: boolean}
        created_at: {type: string, format: date-time}
    CreateImageTask:
      type: object
      required: [type, prompt, aspect_ratio, count]
      properties:
        type:
          type: string
          enum: [TEXT_TO_IMAGE, CHARACTER_REFERENCE]
        prompt: {type: string, maxLength: 1500}
        source_asset_id: {type: string}
        style_id: {type: string}
        aspect_ratio:
          type: string
          enum: ['1:1', '16:9', '9:16', '4:3', '3:4', '3:2', '2:3', '21:9']
        count: {type: integer, minimum: 1, maximum: 4}
        seed: {type: integer, format: int64}
        prompt_optimizer: {type: boolean}
    TaskStatus:
      type: string
      enum: [PENDING, PROCESSING, RETRYING, SUCCEEDED, FAILED, CANCELED]
    ImageTask:
      type: object
      required: [id, type, status, progress, prompt, aspect_ratio, count, credits_reserved, images, created_at]
      properties:
        id: {type: string}
        type: {type: string}
        status:
          $ref: '#/components/schemas/TaskStatus'
        progress: {type: integer, minimum: 0, maximum: 100}
        prompt: {type: string}
        aspect_ratio: {type: string}
        count: {type: integer}
        credits_reserved: {type: integer, format: int64}
        error_code: {type: string}
        error_message: {type: string}
        images:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        created_at: {type: string, format: date-time}
        completed_at: {type: string, format: date-time}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}